recipes-containers: Move docker-app to v0.9
This moves us to docker-app 0.9. There are some big changes in this
version including the fact its not longer able to be used as a
standalone command. This includes a patch to mark the command as
not experimental so that running `docker app` just works. Without the
patch users have to update their docker config.json to enable it.

Signed-off-by: Andy Doan <andy@foundries.io>